Add filepath_test
